Create SearchEngineSelect component
